requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
discord==1.0.1
discord.py==1.7.1
idna==2.10
idna-ssl==1.1.0
lxml==4.6.3
multidict==5.1.0
multitasking==0.0.9
numpy==1.19.5
pandas==1.1.5
pkg-resources==0.0.0
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
yfinance==0.1.59
